fix: address CodeQL findings (zip-slip + implicit narrowing) (#7576)
* DataInstaller: reject archive entries that escape the install dir

CodeQL java/zipslip flagged populateFileAndDirLists at lines 750/753;
the actual sink is createFiles at line 619 (FileOutputStream) and
createDirectories at line 617 (mkdirs), reached via a string-concat
correctFileName that didn't normalise '..' segments.

correctFileName now resolves each entry against its base directory and
rejects any path whose canonical form leaves that base. Both
canonicalisations happen on the resolved File, so symlinks, '.' and
'..' segments are all collapsed before comparison. The two callers
that didn't already throw IOException (checkOverwriteOK and
createDirectories) catch and abort the install with the existing
error-dialog pattern.

A hostile data set containing 'data/../../etc/whatever' would now be
refused before any file is written.

* SkillModifier: make double-to-int truncation explicit

CodeQL java/implicit-cast-in-compound-assignment flagged 12 sites in
this file where a 'double' bonus was accumulated into an 'int' via +=,
which silently inserts (int) at every addition. The intent was always
truncating accumulation -- the function returns Integer and uses
.intValue() for the formula path -- so the warning is purely about
making the cast visible.

Fix: write the cast at every site. No behaviour change; the bytecode
already contained the same i2d/d2i pair.

* DataInstaller: regression test for zip-slip rejection

Reflective unit test so the same class compiles against both the
pre-fix (`private`, no checked exception) and post-fix
(package-private, throws IOException) signatures of correctFileName.
Verified by hand: temporarily reverting DataInstaller to master and
re-running this test yields 1 pass + 2 failures (both '..'-escape
cases are silently accepted under the old logic), and restoring the
fix flips it to 3 passes -- which makes the test a real regression
guard rather than a tautology.

correctFileName goes from `private` to package-private to give the
test in the same package direct access; reflection was needed only
for the cross-state run.

+++ code/src/java/pcgen/core/analysis/SkillModifier.java
@@ -52,48 +52,48 @@ public final class SkillModifier
 		{
 			PCStat stat = statref.get();
 			bonus = aPC.getStatModFor(stat);
-			bonus += aPC.getTotalBonusTo("SKILL", "STAT." + stat.getKeyName());
+			bonus += (int) aPC.getTotalBonusTo("SKILL", "STAT." + stat.getKeyName());
 		}
-		bonus += aPC.getTotalBonusTo("SKILL", keyName);
+		bonus += (int) aPC.getTotalBonusTo("SKILL", keyName);
 
 		// loop through all current skill types checking for boni
 		for (Type singleType : sk.getTrueTypeList(false))
 		{
-			bonus += aPC.getTotalBonusTo("SKILL", "TYPE." + singleType);
+			bonus += (int) aPC.getTotalBonusTo("SKILL", "TYPE." + singleType);
 		}
 
 		// now check for any lists of skills, etc
-		bonus += aPC.getTotalBonusTo("SKILL", "LIST");
+		bonus += (int) aPC.getTotalBonusTo("SKILL", "LIST");
 
 		// now check for ALL
-		bonus += aPC.getTotalBonusTo("SKILL", "ALL");
+		bonus += (int) aPC.getTotalBonusTo("SKILL", "ALL");
 
 		// these next two if-blocks try to get BONUS:[C]CSKILL|TYPE=xxx|y to
 		// function
 		if (aPC.isClassSkill(sk))
 		{
-			bonus += aPC.getTotalBonusTo("CSKILL", keyName);
+			bonus += (int) aPC.getTotalBonusTo("CSKILL", keyName);
 
 			// loop through all current skill types checking for boni
 			for (Type singleType : sk.getTrueTypeList(false))
 			{
-				bonus += aPC.getTotalBonusTo("CSKILL", "TYPE." + singleType);
+				bonus += (int) aPC.getTotalBonusTo("CSKILL", "TYPE." + singleType);
 			}
 
-			bonus += aPC.getTotalBonusTo("CSKILL", "LIST");
+			bonus += (int) aPC.getTotalBonusTo("CSKILL", "LIST");
 		}
 
 		if (!aPC.isClassSkill(sk) && !sk.getSafe(ObjectKey.EXCLUSIVE))
 		{
-			bonus += aPC.getTotalBonusTo("CCSKILL", keyName);
+			bonus += (int) aPC.getTotalBonusTo("CCSKILL", keyName);
 
 			// loop through all current skill types checking for boni
 			for (Type singleType : sk.getTrueTypeList(false))
 			{
-				bonus += aPC.getTotalBonusTo("CCSKILL", "TYPE." + singleType);
+				bonus += (int) aPC.getTotalBonusTo("CCSKILL", "TYPE." + singleType);
 			}
 
-			bonus += aPC.getTotalBonusTo("CCSKILL", "LIST");
+			bonus += (int) aPC.getTotalBonusTo("CCSKILL", "LIST");
 		}
 
 		// the above two if-blocks try to get
@@ -129,7 +129,7 @@ public final class SkillModifier
 				SkillInfoUtilities.getKeyStatList(pc, sk, typeList);
                 for (Type type : typeList)
                 {
-                    statMod += pc.getTotalBonusTo("SKILL", "TYPE." + type);
+                    statMod += (int) pc.getTotalBonusTo("SKILL", "TYPE." + type);
                 }
 			}
 			return statMod;
